# Fix caching being busted by moving .profile update to end of build phase