… tests/snapshots/generate_plan_tests__ruby.snap
… tests/snapshots/generate_plan_tests__ruby_2.snap
… tests/snapshots/generate_plan_tests__ruby_execjs.snap
… tests/snapshots/generate_plan_tests__ruby_local_deps.snap
… tests/snapshots/generate_plan_tests__ruby_rails_postgres.snap
… tests/snapshots/generate_plan_tests__ruby_sinatra.snap
… tests/snapshots/generate_plan_tests__ruby_with_node.snap